Fix caching
Nginx's std cfg only returns 304 with If-Unmodified-Since header, if the date
is exactly the one it expects, not the date the client did its last request.

This makes sense as it is much easier to compare two strings for equality than
to parse the date received from the client and check the ordering with the
server known last_modified value.